==> src/02_merge_files.ipynb <==
import pandas as pd
import os

### Main Code

if __name__ == "__main__":
    directorio = 'D:/Proyecto ironhack/files/datos_coordenadas'
    archivos = [os.path.join(directorio, f) for f in os.listdir(directorio) if f.endswith('.json')]

    dataframes = []

    for archivo in archivos:
        try:
            df = pd.read_json(archivo)
            dataframes.append(df)
        except: 
            pass

    # DataFrame No Normalizado
    df = pd.concat(dataframes, ignore_index=True).drop_duplicates()
    # df.to_csv('D:/Proyecto ironhack/files/spain_dataset_1.csv', index=False)

    # DataFrame Normalizado
    df = pd.json_normalize(df['listings']).drop_duplicates()
    df.to_csv('D:/Proyecto ironhack/files/spain_dataset.csv', index=False)
